# docs/security/openapi.yaml
openapi: 3.0.3
info:
  title: Civis CRM/ERP API
  version: 0.1.0
  description: |
    Generated from the current Next.js App Router endpoints under `app/api/**/route.ts`.
    This spec is intended for Swagger/SwaggerHub documentation and VAPT planning.
servers:
  - url: https://ik-crm-erp.onrender.com
    description: Production
  - url: http://localhost:3000
    description: Local development
...
components:
  securitySchemes:
    nextAuthSession:
      type: apiKey
      in: cookie
      name: next-auth.session-token
      description: NextAuth session cookie (development/non-secure cookie name).
    nextAuthSecureSession:
      type: apiKey
      in: cookie
      name: __Secure-next-auth.session-token
      description: NextAuth session cookie (secure cookie name).
  schemas:
...
    InvoiceInput:
      type: object
      properties:
        invoiceNumber: { type: string }
        clientName: { type: string }
        amount: { type: integer }
        status: { type: string, enum: [DRAFT, SENT, PAID, OVERDUE] }
        dueDate: { type: string, format: date-time }
        issueDate: { type: string, format: date-time }
      required: [invoiceNumber, clientName, amount]
    ExpenseInput:
      type: object
      properties:
        description: { type: string }
        amount: { type: integer }
        category: { type: string }
        date: { type: string, format: date-time }
        status: { type: string, enum: [PENDING, APPROVED, REJECTED] }
        submittedBy: { type: string }
      required: [description, amount]
...
security:
  - nextAuthSession: []
  - nextAuthSecureSession: []
paths:
...
  /api/accounting/invoices:
    get:
      tags: [Accounting]
      summary: List invoices
      responses:
        '200': { description: Invoices list }
    post:
      tags: [Accounting]
      summary: Create invoice
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/InvoiceInput' }
      responses:
        '200': { description: Invoice created }
    patch:
      tags: [Accounting]
      summary: Update invoice
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - $ref: '#/components/schemas/InvoiceInput'
                - type: object
                  properties:
                    id: { type: string }
                  required: [id]
      responses:
        '200': { description: Invoice updated }
    delete:
      tags: [Accounting]
      summary: Delete invoice
      parameters:
        - name: id
          in: query
          required: true
          schema: { type: string }
      responses:
        '200': { description: Invoice deleted }
  /api/accounting/expenses:
    get:
      tags: [Accounting]
      summary: List expenses
      responses:
        '200': { description: Expenses list }
    post:
      tags: [Accounting]
      summary: Create expense
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ExpenseInput' }
      responses:
        '200': { description: Expense created }
    patch:
      tags: [Accounting]
      summary: Update expense
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - $ref: '#/components/schemas/ExpenseInput'
                - type: object
                  properties:
                    id: { type: string }
                  required: [id]
      responses:
        '200': { description: Expense updated }
    delete:
      tags: [Accounting]
      summary: Delete expense
      parameters:
        - name: id
          in: query
          required: true
          schema: { type: string }
      responses:
        '200': { description: Expense deleted }
...